typographyStyles file created, getUser fn improved, CustomAppBar moved to CustomComponents folder
