+++ src/Styles/typographyStyles.js
@@ -1,0 +1,7 @@
+export const paragraphText40 = {
+  fontSize: 40,
+  textAlign: 'center',
+  marginTop: 20,
+};
+
+export const paragraphText45 = { fontSize: 45, textAlign: 'center' };
